Update layout and UI elements for better readability and design consistency
The NavigationBar in layout.svelte was commented out to assess its necessity and effect on the layout, which will help in deciding whether or not to altogether remove it in the future. The LargeWorkCard.svelte underwent several changes primarily aimed at improving readability by manipulating background image opacity, adding shadow effect and tweaking its positioning and padding. These improvements aim to make the card details more prominent and the card as a whole more in line with the web app's existing design elements.