Add WIP component to concept page
A "Work In Progress" (WIP) component was imported and added to the concept page. This action was taken to replace a block of placeholder text and images. Currently, the content for this page is not ready. Adding the WIP component serves as an indication that content will be populated in the future.