Update layout and styles of multiple sections and cards
Made adjustments to the layout and styles of various elements across the website, including the About, Contact, and Work sections, and the About and Work cards. The changes include altering padding, text sizes, class names, and div structures for improved readability and aesthetics. Also, the handling of pictures in the About cards was modified to allow them to be full images. These changes help improve user experience and the overall look and feel of the website.